+++ backend/common/util.go
@@ -10,8 +10,11 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
+	"strconv"
 	"strings"
 	"time"
+
+	"github.com/pkg/errors"
 )
 
 const (
@@ -184,4 +187,31 @@ func Unobfuscate(dst, seed string) (string, error) {
 		unobfuscated[i] = b ^ seedBytes[i%len(seedBytes)]
 	}
 	return string(unobfuscated), nil
+}
+
+// NormalizeExternalURL will format the external url.
+func NormalizeExternalURL(url string) (string, error) {
+	r := strings.TrimSpace(url)
+	r = strings.TrimSuffix(r, "/")
+	if !HasPrefixes(r, "http://", "https://") {
+		return "", errors.Errorf("%s must start with http:// or https://", url)
+	}
+	parts := strings.Split(r, ":")
+	if len(parts) > 3 {
+		return "", errors.Errorf("%s malformed", url)
+	}
+	if len(parts) == 3 {
+		port, err := strconv.Atoi(parts[2])
+		if err != nil {
+			return "", errors.Errorf("%s has non integer port", url)
+		}
+		// The external URL is used as the redirectURL in the get token process of OAuth, and the
+		// RedirectURL needs to be consistent with the RedirectURL in the get code process.
+		// The frontend gets it through window.location.origin in the get code
+		// process, so port 80/443 need to be cropped.
+		if port == 80 || port == 443 {
+			r = strings.Join(parts[0:2], ":")
+		}
+	}
+	return r, nil
 }
